requirements.txt
chumpy==0.70
easydict==1.10
imageio==2.27.0
joblib==1.2.0
matplotlib==3.5.3
opencv-python==4.7.0.72
pycocotools==2.0.6
pyglet==1.5.0
pyrender==0.1.45
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
smplx==0.1.28
tqdm==4.65.0
trimesh==3.21.5
typing_extensions==4.5.0
pyyaml
torchgeometry==0.1.2